Add validations to the fields of the create and edit event forms
Obviously this validation is only aesthetic in the frontend to guide the user, in the backend there is all this validation again.
To guide the user, the incorrect fields are colored red and a <div> is shown with the message that what is wrong. The "Submit" button is also disabled while the form is invalid
- All fields are required
- The date is checked to be after now
- The map that is an iframe
- The minimum price is 0
- The capacity is at least 1